Play nice with macOS
## Changes Description

This commit makes the "Ensure sudo is installed" to not be executed on
macOS (if you have Homebrew installed, this task fails because `brew`
cannot be executed as `root`).

It also allows the user to change the group owner used to create and
verify files. Here, if the user does not change the default value, we
check if the distribution we're running in is a MacOSX, and if it is
we use `wheel` instead of `root`.